- Make bitsearch automatically use the correct function for the type - Put const on the right where it belongs - Add q suffix on lidtq and lgdtq - Add debug exception handler (which just returns) to handle spurious debug exceptions when debugging stress - Update ISR handler code to preserve stack trace - Replace physical memory allocator with a new implementation that does reference counting - Update disk read_blocks and write_blocks return types - Implement IDE driver - Change unnecessary use of RIP-relative addressing - Add a few extern "C" to make debugging simpler - Implement enable_if, conditional, integral_constant, true_type, false_type - Add debug message for filesystem mount - Add temporary handling of clipped lines (discard whole line if clipping is needed) - Add mphysranges_split to post-process a physical range list to ensure it doesn't cross a power of two boundary - Add a way to reserve address space and alias ranges of memory (for device I/O) - Add operator delete[] - Increase maximum spin count in threadsync.cc
-

+++ kernel/arch/x86_64/device/ahci.cc
@@ -1190,7 +1190,7 @@ struct ahci_if_t : public storage_if_base_t {
 struct ahci_dev_t : public storage_dev_base_t {
     STORAGE_DEV_IMPL
 
-    int io(void *data, uint64_t count,
+    int io(void *data, int64_t count,
            uint64_t lba, int is_read);
 
     ahci_if_t *if_;
@@ -1965,7 +1965,7 @@ static void ahci_async_complete(int error,
         condvar_wake_one(&state->done_cond);
 }
 
-int ahci_dev_t::io(void *data, uint64_t count,
+int ahci_dev_t::io(void *data, int64_t count,
                   uint64_t lba, int is_read)
 {
     ahci_blocking_io_t block_state;
@@ -1997,15 +1997,15 @@ int ahci_dev_t::io(void *data, uint64_t count,
     return block_state.err;
 }
 
-int ahci_dev_t::read_blocks(
-        void *data, uint64_t count,
+int64_t ahci_dev_t::read_blocks(
+        void *data, int64_t count,
         uint64_t lba)
 {
     return io(data, count, lba, 1);
 }
 
-int ahci_dev_t::write_blocks(
-        void const *data, uint64_t count,
+int64_t ahci_dev_t::write_blocks(
+        void const *data, int64_t count,
         uint64_t lba)
 {
     return io((void*)data, count, lba, 0);
